requirements.txt
selenium==4.35.0
Flask==3.1.2
python-dotenv==1.1.1
colorlog==6.9.0
SQLAlchemy==2.0.43
PyMySQL==1.1.2

